+++ examples/data-sources/nxos_subinterface_vrf/data-source.tf
@@ -1,3 +1,0 @@
-data "nxos_subinterface_vrf" "example" {
-  interface_id = "eth1/10.124"
-}