+++ examples/resources/nxos_subinterface_vrf/import-by-identity.tf
@@ -1,6 +1,0 @@
-import {
-  to = nxos_subinterface_vrf.example
-  identity = {
-    "interface_id" : "<interface_id>"
-  }
-}