+++ internal/provider/data_source_nxos_subinterface_test.go
@@ -40,6 +40,7 @@ func TestAccDataSourceNxosSubinterface(t *testing.T) {
 	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_subinterface.test", "link_logging", "enable"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_subinterface.test", "medium", "broadcast"))
 	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_subinterface.test", "mtu", "1500"))
+	checks = append(checks, resource.TestCheckResourceAttr("data.nxos_subinterface.test", "vrf_dn", "sys/inst-VRF123"))
 	resource.Test(t, resource.TestCase{
 		PreCheck:                 func() { testAccPreCheck(t) },
 		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
@@ -81,6 +82,7 @@ func testAccDataSourceNxosSubinterfaceConfig() string {
 	config += `	link_logging = "enable"` + "\n"
 	config += `	medium = "broadcast"` + "\n"
 	config += `	mtu = 1500` + "\n"
+	config += `	vrf_dn = "sys/inst-VRF123"` + "\n"
 	config += `	depends_on = [nxos_rest.PreReq0, ]` + "\n"
 	config += `}` + "\n"